Move thread_pool.hpp to src/ and consolidate wrapper
- Rename include/packingsolver/algorithms/thread_pool.hpp →
  src/algorithms/thread_pool.hpp (internal-only header; no public API)
- Move wrapper_impl/wrapper templates from common.hpp into thread_pool.hpp
  so all threading helpers live in one place
- Update include paths in all optimize.cpp files and column_generation_2.cpp
- Refactor column_generation_2.cpp to use run(tasks, true) instead of
  manual std::thread management; fixes a latent data race where both
  threads shared a single exception_ptr
- Update thread_pool.hpp header comment to document all three helpers
- Start memory monitor thread regardless of parallel mode
