+++ src/onedimensional/optimize.cpp
@@ -6,7 +6,7 @@
 #include "algorithms/dichotomic_search.hpp"
 #include "algorithms/sequential_value_correction.hpp"
 #include "algorithms/column_generation.hpp"
-#include "packingsolver/algorithms/thread_pool.hpp"
+#include "algorithms/thread_pool.hpp"
 
 #include "treesearchsolver/iterative_beam_search_2.hpp"
